requirements.txt
django-debug-toolbar
Django==3.0.1
django-model-utils==3.2.0
django-multiselectfield==0.1.8
pytz==2019.2
sqlparse==0.3.0
